+++ aeon/transformations/collection/convolution_based/_minirocket.py
