+++ aeon/transformations/collection/convolution_based/_minirocket_multivariate_variable.py
