+++ aeon/transformations/collection/convolution_based/_multirocket.py
