+++ aeon/transformations/collection/convolution_based/_multirocket_multivariate.py
