+++ aeon/transformations/collection/convolution_based/_rocket.py
